movable Interface 생성
Monster를 타입 별로 Flight, Runner, Esper로 나누었고
몬스터들이 Movable을 상속받아 각 클래스에서 move를 재정의하게 만들었다
