--- src/com/hsik0225/Fly.java
@@ -1,4 +1,0 @@
-package com.hsik0225;
-
-public class Fly {
-}
