+++ src/com/hsik0225/movable.java
@@ -1,0 +1,5 @@
+package com.hsik0225;
+
+public interface movable {
+    public String move(int randomValue);
+}
